Multiple changes
- Moves instantiate code to its own instantiate.{hpp,cpp} files.
- Updates instantiate function to optionally take an options parameter
  that can specify the available descriptors as well as environment.
- Updates shell to instantiate defined systems.
- Updates shell with support for waiting for instantiated systems.
- Updates shell to use more libedit features.
- Adds connect_with_user convenience function.
